Fixed non-working tests and added some more
The original test file had a failing test in line 34. I had to change a bit how the Boggle class works so I could create a special board and see if the Boggle class was working correctly.

I also added a requirements-dev.txt with ipdb and pytest, just because they allow me to test and find bugs easier.